+++ template/src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import logo from './logo.svg';
 import './App.css';
 import { useSelector, useDispatch } from 'react-redux';
-import { selectors, actions } from './ducks/counter';
+import { selectors, actions } from './features/counter';
 
 function App() {
   const count = useSelector(selectors.selectCount);
