+++ template/src/features/counter.js
